Move auth related components to specific package
