+++ src/app/auth/signup/signup.component.ts
@@ -3,8 +3,8 @@ import { Router } from '@angular/router';
 import { FormGroup, FormBuilder, Validators } from '@angular/forms';
 import { ViewEncapsulation} from '@angular/core';
 
-import { AuthenticationService } from '../services/authentication.service';
-import { CustomValidators } from '../shared/custom-validators';
+import { AuthenticationService } from '../../services/authentication.service';
+import { CustomValidators } from '../../shared/custom-validators';
 
 @Component({
   selector: 'app-signup',
