+++ src/app/auth/update-pass/update-pass.component.scss
