+++ src/app/home/home.component.scss
@@ -1,4 +1,4 @@
-@import "../signup/signup.component.scss";
+@import "../auth/signup/signup.component.scss";
 @import url('https://fonts.googleapis.com/css?family=Open+Sans');
 
 h3 {
